Renamed core package to aboutthisapp
