+++ build.gradle
@@ -2,15 +2,17 @@
 
 buildscript {
     ext.kotlin_version = '1.4.30'
+    ext.junit = "5.6.2"
     repositories {
         google()
         jcenter()
         
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.0.2'
+        classpath 'com.android.tools.build:gradle:4.1.3'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath "com.github.dcendents:android-maven-gradle-plugin:2.1"
+        classpath "de.mannodermaus.gradle.plugins:android-junit5:1.7.1.1"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
@@ -21,6 +23,17 @@ allprojects {
         google()
         jcenter()
         maven { url 'https://jitpack.io' }
+    }
+
+    tasks.withType(Test) {
+        testLogging {
+            exceptionFormat "full"
+            showCauses true
+            showExceptions true
+            showStackTraces true
+            showStandardStreams true
+            events = ["passed", "skipped", "failed", "standardOut", "standardError"]
+        }
     }
 }
