+++ library/src/main/java/tmg/aboutthisapp/utils/Constants.kt
@@ -1,3 +1,3 @@
-package tmg.components.utils
+package tmg.aboutthisapp.utils
 
 const val marketUri: String = "https://play.google.com/store/apps/details?id=%s"
